Extract common code of sync rule wizards in generic component
